service/backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP client for downloads
aiohttp==3.9.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# System monitoring
psutil==5.9.8

# Google AI (Gemini)
google-generativeai>=0.7.0
